# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Al-Khatib Contracting
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface InquiryRequest {
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  serviceType: string;
  serviceTitle?: string;
  message: string;
  address?: string;
}

export interface InquiryResponse {
  id: number;
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  serviceType: string;
  serviceTitle?: string;
  message: string;
  address?: string;
  createdAt: string;
  status: string;
}

export type ServiceRequestCategory =
  (typeof ServiceRequestCategory)[keyof typeof ServiceRequestCategory];

export const ServiceRequestCategory = {
  interior: "interior",
  exterior: "exterior",
  maintenance: "maintenance",
} as const;

export interface ServiceRequest {
  title: string;
  description: string;
  category: ServiceRequestCategory;
  icon?: string;
  images?: string[];
  features?: string[];
  isActive?: boolean;
  sortOrder?: number;
}

export interface ServiceResponse {
  id: number;
  title: string;
  description: string;
  category: string;
  icon?: string;
  images?: string[];
  features?: string[];
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectRequest {
  title: string;
  description: string;
  category: string;
  location?: string;
  completionDate?: string;
  area?: string;
  images?: string[];
  features?: string[];
  isActive?: boolean;
  isFeatured?: boolean;
  sortOrder?: number;
}

export interface ProjectResponse {
  id: number;
  title: string;
  description: string;
  category: string;
  location?: string;
  completionDate?: string;
  area?: string;
  images?: string[];
  features?: string[];
  isActive: boolean;
  isFeatured: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}