# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API for Al-Khatib Contracting
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: inquiries
    description: Customer inquiry operations
  - name: services
    description: Services management
  - name: projects
    description: Projects management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /inquiries:
    post:
      operationId: submitInquiry
      tags: [inquiries]
      summary: Submit a customer inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryRequest"
      responses:
        "201":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InquiryResponse"
    get:
      operationId: getInquiries
      tags: [inquiries]
      summary: Get all inquiries
      responses:
        "200":
          description: List of inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InquiryResponse"

  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: Get all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceResponse"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"

  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequest"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: Get all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    InquiryRequest:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        serviceType:
          type: string
        serviceTitle:
          type: string
        message:
          type: string
        address:
          type: string
      required:
        - customerName
        - customerPhone
        - serviceType
        - message

    InquiryResponse:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        serviceType:
          type: string
        serviceTitle:
          type: string
        message:
          type: string
        address:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          type: string
      required:
        - id
        - customerName
        - customerPhone
        - serviceType
        - message
        - createdAt
        - status

    ServiceRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [interior, exterior, maintenance]
        icon:
          type: string
        images:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - title
        - description
        - category

    ServiceResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
        images:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - isActive
        - sortOrder
        - createdAt
        - updatedAt

    ProjectRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        completionDate:
          type: string
        area:
          type: string
        images:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
      required:
        - title
        - description
        - category

    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        completionDate:
          type: string
        area:
          type: string
        images:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - isActive
        - isFeatured
        - sortOrder
        - createdAt
        - updatedAt