# lib/api-zod/src/generated/types/deleteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Al-Khatib Contracting
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteResponse {
  success: boolean;
  message: string;
}