# lib/api-zod/src/generated/types/projectResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Al-Khatib Contracting
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectResponse {
  id: number;
  title: string;
  description: string;
  category: string;
  location?: string;
  completionDate?: string;
  area?: string;
  images?: string[];
  features?: string[];
  isActive: boolean;
  isFeatured: boolean;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}