# lib/api-zod/src/generated/types/serviceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Al-Khatib Contracting
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceRequestCategory } from "./serviceRequestCategory";

export interface ServiceRequest {
  title: string;
  description: string;
  category: ServiceRequestCategory;
  icon?: string;
  images?: string[];
  features?: string[];
  isActive?: boolean;
  sortOrder?: number;
}