# lib/api-zod/src/generated/types/serviceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Al-Khatib Contracting
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceResponse {
  id: number;
  title: string;
  description: string;
  category: string;
  icon?: string;
  images?: string[];
  features?: string[];
  isActive: boolean;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}